refactor: split validator tests into separate files
ref: https://github.com/OpenPathfinder/visionBoard/issues/190
